+++ src/Makefile.am
@@ -255,6 +255,7 @@ libbitcoin_server_a_SOURCES = \
   init.cpp \
   leveldbwrapper.cpp \
   main.cpp \
+  txBaseMsgProcessor.cpp \
   merkleblock.cpp \
   metrics.cpp \
   miner.cpp \
