+++ src/gtest/test_processTxBaseMsg.cpp
@@ -2,6 +2,7 @@
 #include <gtest/tx_creation_utils.h>
 #include <vector>
 
+#include <txBaseMsgProcessor.h>
 #include <main.h>
 #include <protocol.h>
 #include <txdb.h>
@@ -174,10 +175,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, ValidTxIsRelayed)
     ASSERT_TRUE(mapRelay.count(CInv(MSG_TX, validTx.GetHash())) == 0);
 
     FakeNode sourceNode{};
-    addTxBaseMsgToProcess(validTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(validTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, validTx.GetHash())) != 0);
@@ -191,10 +192,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, ValidTxIsRecordedAsKnown)
     ASSERT_FALSE(AlreadyHave(CInv(MSG_TX, validTx.GetHash())));
 
     FakeNode sourceNode{};
-    addTxBaseMsgToProcess(validTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(validTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(AlreadyHave(CInv(MSG_TX, validTx.GetHash())));
@@ -212,10 +213,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, RetransmittedValidTxFromNonWhitelistedNodeIsNo
     ASSERT_TRUE(mapRelay.count(CInv(MSG_TX, validTx.GetHash())) == 0);
 
     FakeNode sourceNode{};
-    addTxBaseMsgToProcess(validTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(validTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, validTx.GetHash())) == 0);
@@ -234,10 +235,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, RetransmittedValidTxFromWhitelistedNodeIsRelay
 
     FakeNode sourceNode{};
     sourceNode.fWhitelisted = true;
-    addTxBaseMsgToProcess(validTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(validTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, validTx.GetHash())) != 0);
@@ -251,10 +252,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, NoRejectMsgIsSentForValidATx)
 
     FakeNode sourceNode{};
     ASSERT_TRUE(sourceNode.commandInvoked.empty());
-    addTxBaseMsgToProcess(validTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(validTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(sourceNode.commandInvoked.empty());
@@ -271,10 +272,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, MissingInputsNonJoinsplitTxIsNotRelayed)
     ASSERT_TRUE(mapRelay.count(CInv(MSG_TX, orphanNonJoinsplitTx.GetHash())) == 0);
 
     FakeNode sourceNode{};
-    addTxBaseMsgToProcess(orphanNonJoinsplitTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(orphanNonJoinsplitTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, orphanNonJoinsplitTx.GetHash())) == 0);
@@ -288,10 +289,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, MissingInputsNonJoinsplitTxIsRecordedAsKnown)
     ASSERT_FALSE(AlreadyHave(CInv(MSG_TX, orphanNonJoinsplitTx.GetHash())));
 
     FakeNode sourceNode{};
-    addTxBaseMsgToProcess(orphanNonJoinsplitTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(orphanNonJoinsplitTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(AlreadyHave(CInv(MSG_TX, orphanNonJoinsplitTx.GetHash())));
@@ -310,10 +311,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, RetransmittedMissingInputsNonJoinsplitTxFromNo
     ASSERT_TRUE(mapRelay.count(CInv(MSG_TX, orphanNonJoinsplitTx.GetHash())) == 0);
 
     FakeNode sourceNode{};
-    addTxBaseMsgToProcess(orphanNonJoinsplitTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(orphanNonJoinsplitTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, orphanNonJoinsplitTx.GetHash())) == 0);
@@ -333,10 +334,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, RetransmittedMissingInputsNonJoinsplitTxFromWh
 
     FakeNode sourceNode{};
     sourceNode.fWhitelisted = true;
-    addTxBaseMsgToProcess(orphanNonJoinsplitTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(orphanNonJoinsplitTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, orphanNonJoinsplitTx.GetHash())) != 0);
@@ -350,10 +351,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, NoRejectMsgIsSentForMissingInputsNonJoinsplitT
 
     FakeNode sourceNode{};
     ASSERT_TRUE(sourceNode.commandInvoked.empty());
-    addTxBaseMsgToProcess(orphanNonJoinsplitTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(orphanNonJoinsplitTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(sourceNode.commandInvoked.empty());
@@ -370,10 +371,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, MissingInputsJoinsplitTxIsNotRelayed)
     ASSERT_TRUE(mapRelay.count(CInv(MSG_TX, orphanJoinsplitTx.GetHash())) == 0);
 
     FakeNode sourceNode{};
-    addTxBaseMsgToProcess(orphanJoinsplitTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(orphanJoinsplitTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, orphanJoinsplitTx.GetHash())) == 0);
@@ -389,10 +390,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, MissingInputsJoinsplitTxFromWhitelistedPeerIsR
 
     FakeNode sourceNode{};
     sourceNode.fWhitelisted = true;
-    addTxBaseMsgToProcess(orphanJoinsplitTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(orphanJoinsplitTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, orphanJoinsplitTx.GetHash())) != 0);
@@ -407,10 +408,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, MissingInputsJoinsplitTxIsRecordedAsKnown)
     ASSERT_FALSE(AlreadyHave(CInv(MSG_TX, orphanJoinsplitTx.GetHash())));
 
     FakeNode sourceNode{};
-    addTxBaseMsgToProcess(orphanJoinsplitTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(orphanJoinsplitTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(AlreadyHave(CInv(MSG_TX, orphanJoinsplitTx.GetHash())));
@@ -430,10 +431,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, RetransmittedMissingInputsJoinsplitTxFromNonWh
     ASSERT_TRUE(mapRelay.count(CInv(MSG_TX, orphanJoinsplitTx.GetHash())) == 0);
 
     FakeNode sourceNode{};
-    addTxBaseMsgToProcess(orphanJoinsplitTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(orphanJoinsplitTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, orphanJoinsplitTx.GetHash())) == 0);
@@ -454,10 +455,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, RetransmittedMissingInputsJoinsplitTxFromWhite
 
     FakeNode sourceNode{};
     sourceNode.fWhitelisted = true;
-    addTxBaseMsgToProcess(orphanJoinsplitTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(orphanJoinsplitTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, orphanJoinsplitTx.GetHash())) != 0);
@@ -472,10 +473,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, NoRejectMsgIsSentForMissingInputsJoinsplitTx)
 
     FakeNode sourceNode{};
     ASSERT_TRUE(sourceNode.commandInvoked.empty());
-    addTxBaseMsgToProcess(orphanJoinsplitTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(orphanJoinsplitTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(sourceNode.commandInvoked.empty());
@@ -494,10 +495,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, InvalidZeroDosTxIsNotRelayed)
     ASSERT_TRUE(mapRelay.count(CInv(MSG_TX, invalidZeroDoSTx.GetHash())) == 0);
 
     FakeNode sourceNode{};
-    addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, invalidZeroDoSTx.GetHash())) == 0);
@@ -514,10 +515,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, InvalidZeroDosTxFromWhitelistedPeerIsRelayed)
 
     FakeNode sourceNode{};
     sourceNode.fWhitelisted = true;
-    addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, invalidZeroDoSTx.GetHash())) != 0);
@@ -533,10 +534,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, InvalidZeroDosTxIsRecordedAsKnown)
     ASSERT_FALSE(AlreadyHave(CInv(MSG_TX, invalidZeroDoSTx.GetHash())));
 
     FakeNode sourceNode{};
-    addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(AlreadyHave(CInv(MSG_TX, invalidZeroDoSTx.GetHash())));
@@ -555,13 +556,13 @@ TEST_F(ProcessTxBaseMsgTestSuite, RetransmittedInvalidZeroDosTxFromNonWhiteliste
     FakeNode sourceNode{};
 
     // process first time
-    addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
     ASSERT_TRUE(AlreadyHave(CInv(MSG_TX, invalidZeroDoSTx.GetHash())));
 
     //test
-    addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, invalidZeroDoSTx.GetHash())) == 0);
@@ -580,16 +581,16 @@ TEST_F(ProcessTxBaseMsgTestSuite, RetransmittedInvalidZeroDosTxFromWhitelistedNo
     FakeNode sourceNode{};
 
     // process first time
-    addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
     ASSERT_TRUE(AlreadyHave(CInv(MSG_TX, invalidZeroDoSTx.GetHash())));
 
     // Set whitelist and retry
     sourceNode.fWhitelisted = true;
 
     //test
-    addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, invalidZeroDoSTx.GetHash())) != 0);
@@ -604,11 +605,11 @@ TEST_F(ProcessTxBaseMsgTestSuite, RejectMsgIsSentForInvalidZeroDosTx)
     theFake.DosLevelIfInvalid = 0;
 
     FakeNode sourceNode{};
-    addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(invalidZeroDoSTx, &sourceNode);
     ASSERT_TRUE(sourceNode.commandInvoked.empty());
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(sourceNode.commandInvoked == std::string("rejecttx"));
@@ -627,10 +628,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, InvalidHighDosTxIsNotRelayed)
     ASSERT_TRUE(mapRelay.count(CInv(MSG_TX, invalidHighDoSTx.GetHash())) == 0);
 
     FakeNode sourceNode{};
-    addTxBaseMsgToProcess(invalidHighDoSTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(invalidHighDoSTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, invalidHighDoSTx.GetHash())) == 0);
@@ -647,10 +648,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, InvalidHighDosTxFromWhitelistedPeerIsNotRelaye
 
     FakeNode sourceNode{};
     sourceNode.fWhitelisted = true;
-    addTxBaseMsgToProcess(invalidHighDoSTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(invalidHighDoSTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, invalidHighDoSTx.GetHash())) == 0);
@@ -666,10 +667,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, InvalidHighDosTxIsRecordedAsKnown)
     ASSERT_FALSE(AlreadyHave(CInv(MSG_TX, invalidHighDoSTx.GetHash())));
 
     FakeNode sourceNode{};
-    addTxBaseMsgToProcess(invalidHighDoSTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(invalidHighDoSTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(AlreadyHave(CInv(MSG_TX, invalidHighDoSTx.GetHash())));
@@ -685,10 +686,10 @@ TEST_F(ProcessTxBaseMsgTestSuite, RejectMsgIsSentForInvalidHighDosTx)
 
     FakeNode sourceNode{};
     ASSERT_TRUE(sourceNode.commandInvoked.empty());
-    addTxBaseMsgToProcess(invalidHighDoSTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(invalidHighDoSTx, &sourceNode);
 
     //test
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(sourceNode.commandInvoked == std::string("rejecttx"));
@@ -717,17 +718,17 @@ TEST_F(ProcessTxBaseMsgTestSuite, OrphanTxesTurningValidAreRelayed)
     FakeNode sourceNode{};
 
     // Orphan txes are inserted first and not relayed
-    addTxBaseMsgToProcess(orphanTx_1, &sourceNode);
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(orphanTx_1, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
     ASSERT_TRUE(mapRelay.count(CInv(MSG_TX, orphanTx_1.GetHash())) == 0);
 
-    addTxBaseMsgToProcess(orphanTx_2, &sourceNode);
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(orphanTx_2, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
     ASSERT_TRUE(mapRelay.count(CInv(MSG_TX, orphanTx_2.GetHash())) == 0);
 
     //test finally the valid parent tx is processed
-    addTxBaseMsgToProcess(validTx, &sourceNode);
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(validTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, validTx.GetHash())) != 0);
@@ -751,13 +752,13 @@ TEST_F(ProcessTxBaseMsgTestSuite, OrphanTxesStayingOrphanAreNotRelayed)
     FakeNode sourceNode{};
 
     // Orphan txes are inserted first and not relayed
-    addTxBaseMsgToProcess(orphanTx_1, &sourceNode);
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(orphanTx_1, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
     ASSERT_TRUE(mapRelay.count(CInv(MSG_TX, orphanTx_1.GetHash())) == 0);
 
     //test finally the valid parent tx is processed
-    addTxBaseMsgToProcess(validTx, &sourceNode);
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(validTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, validTx.GetHash())) != 0);
@@ -779,8 +780,8 @@ TEST_F(ProcessTxBaseMsgTestSuite, OrphanTxesTurningZeroDosInvalidAreNotRelayed)
     FakeNode sourceNode{};
 
     // Orphan tx is inserted first and not relayed, classified as missing input
-    addTxBaseMsgToProcess(invalidMissingInputsTx, &sourceNode);
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(invalidMissingInputsTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
     ASSERT_TRUE(mapRelay.count(CInv(MSG_TX, invalidMissingInputsTx.GetHash())) == 0);
 
     // then orphan is marked as invalid
@@ -788,8 +789,8 @@ TEST_F(ProcessTxBaseMsgTestSuite, OrphanTxesTurningZeroDosInvalidAreNotRelayed)
     theFake.markTxAsInvalid(invalidMissingInputsTx.GetHash());
 
     //test finally the valid parent tx is processed
-    addTxBaseMsgToProcess(validTx, &sourceNode);
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(validTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, validTx.GetHash())) != 0);
@@ -811,8 +812,8 @@ TEST_F(ProcessTxBaseMsgTestSuite, OrphanTxesTurningHighDosInvalidAreNotRelayed)
     FakeNode sourceNode{};
 
     // Orphan tx is inserted first and not relayed, classified as missing input
-    addTxBaseMsgToProcess(invalidMissingInputsTx, &sourceNode);
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(invalidMissingInputsTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
     ASSERT_TRUE(mapRelay.count(CInv(MSG_TX, invalidMissingInputsTx.GetHash())) == 0);
 
     // then orphan is marked as invalid
@@ -820,8 +821,8 @@ TEST_F(ProcessTxBaseMsgTestSuite, OrphanTxesTurningHighDosInvalidAreNotRelayed)
     theFake.markTxAsInvalid(invalidMissingInputsTx.GetHash());
 
     //test finally the valid parent tx is processed
-    addTxBaseMsgToProcess(validTx, &sourceNode);
-    ProcessTxBaseMsg(fakedMemProcessor);
+    TxBaseMsgProcessor::getProcessor().addTxBaseMsgToProcess(validTx, &sourceNode);
+    TxBaseMsgProcessor::getProcessor().ProcessTxBaseMsg(fakedMemProcessor);
 
     //checks
     EXPECT_TRUE(mapRelay.count(CInv(MSG_TX, validTx.GetHash())) != 0);
